refactor prompt context helpers
